Add support for if statements and comparison operators
Introduces parsing, lexing, and code generation for 'if' statements, as well as '==' and '!=' comparison operators. Updates the AST to include NODE_IF and NODE_BLOCK, extends the lexer and parser to recognize new tokens, and modifies codegen to emit conditional branching. The test generator is updated to produce a sample program using these new features.